langgraph/requirements.txt
langchain==0.3.18
langchain-core==0.3.45
langchain-community==0.3.17
langchain-openai==0.3.5
langchain-google-genai==2.0.9
langgraph==0.3.21
langsmith==0.3.7
pydantic==2.10.6
python-dotenv==1.0.1
openai==1.62.0
httpx==0.28.1
graphviz 
pydot
faiss-cpu==1.9.0
httpx==0.27.2
tenacity==9.1.2
python-dotenv==1.0.1
